Revert "Revert "Bloqade-analog migration""
