+++ pyproject.toml
@@ -43,7 +43,7 @@ Discord = "https://discord.gg/TPBU2sa8Et"
 
 [project.optional-dependencies]
 azure = ["azure-quantum>=2.0,<2.3", "azure-storage-blob>=12.20,<13.0", "azure-identity>=1.17,<2.0"]
-bloqade = ["bloqade>=0.15.13,<=0.15.14"]
+bloqade = ["bloqade-analog>=0.16.1,<0.17.0"]
 braket = ["amazon-braket-sdk>=1.83.0,<1.91.0", "pytket-braket>=0.30,<0.40"]
 cirq = ["cirq-core>=1.3,<1.5", "cirq-ionq>=1.3,<1.5", "ply>=3.6", "attrs>=21.3.0"]
 cudaq = ["cudaq>=0.9.0"]
@@ -55,10 +55,8 @@ pyquil = ["pyquil>=4.4", "qcs-sdk-python>=0.21.12"]
 pytket = ["pytket>=1.31"]
 qir = ["qbraid-qir>=0.2.0,<=0.3.1", "qbraid-core[runner]>=0.1.29", "qiskit-qir>=0.5.0"]
 qiskit = ["qiskit>=1.0,<1.5", "qiskit-ibm-runtime>=0.25.0,<0.37", "qiskit-qasm3-import>=0.5.1", "packaging>=20.0"]
-quera = ["flair-visual[vis]>=0.5.3"]
+quera = ["flair-visual[vis]>=0.5.3", "pandas"]
 visualization = ["ipython", "matplotlib", "pylatexenc", "ipympl"]
-test = ["pytest", "pytest-cov", "sympy", "packaging"]
-lint = ["black", "isort", "pylint", "qbraid-cli>=0.10.0"]
 docs = ["sphinx>=7.2,<8.3", "sphinx-autodoc-typehints>=1.24,<3.2", "sphinx-rtd-theme>=1.3,<3.1", "docutils<0.22", "sphinx-copybutton"]
 
 [project.entry-points."qbraid.programs"]
