+++ requirements-cudaq.txt
@@ -1,0 +1,1 @@
+cudaq>=0.9.0; python_version < "3.13"
