+++ requirements-dev.txt
@@ -6,14 +6,17 @@ pyquil>=4.4; python_version < "3.13"
 qcs-sdk-python>=0.21.12,<0.21.13; python_version < "3.13"
 pytket>=1.31
 qiskit>=1.0,<1.5
-cudaq>=0.9.0; python_version < "3.13"
+
+# quera dependencies
+bloqade-analog>=0.16.2,<0.17.0; python_version < "3.13"
+flair-visual>=0.5.3; python_version < "3.13"
+pandas
 
 # transpiler extras
 ply>=3.6
 cirq-ionq>=1.3,<1.5
 qbraid-qir>=0.2.0,<=0.3.1
 pytket-braket>=0.35.1,<0.40.0
-bloqade>=0.15.13,<=0.15.14; python_version < "3.13"
 qiskit-qasm3-import>=0.5.1
 qiskit-aer>=0.15.0; python_version < "3.13"
 qiskit-qir
@@ -29,12 +32,10 @@ pyqubo>=1.4.0
 
 # visualization
 ipython
-matplotlib>=3.3
-pandas>=2.2
-flair-visual>=0.5.3; python_version < "3.13"
+matplotlib>=3.3,<3.8
 
-# testing
+# testing dependencies
 sympy
+packaging>=20.0
 pytest
 pytest-cov
-packaging>=20.0
