+++ tox.ini
@@ -2,6 +2,7 @@
 minversion = 4.2.0
 envlist =
     unit-tests
+    cudaq-tests
     docs
     linters
     format-check
@@ -28,10 +29,30 @@ pass_env =
     QBRAID_RUN_REMOTE_TESTS
 
 [testenv:unit-tests]
-description = Run pytests and generate coverage report.
+description = Run pytests for main package and generate coverage report.
 commands =
     python3 -c "import qbraid; qbraid.about()"
-    pytest tests --cov=qbraid --cov-config=pyproject.toml --cov-report=term --cov-report=xml {posargs}
+    pytest tests \
+        --cov=qbraid \
+        --cov-config=pyproject.toml \
+        --cov-report=term \
+        --cov-report=xml \
+        {posargs}
+
+[testenv:cudaq-tests]
+description = Run pytests for CUDA-Q intergration and generate coverage report.
+deps += -r{toxinidir}/requirements-cudaq.txt
+commands =
+    python3 -c "import qbraid; qbraid.about()"
+    pytest \
+        tests/programs/test_program_cudaq.py \
+        tests/transpiler/qasm/test_qasm_to_cudaq.py \
+        --cov=qbraid \
+        --cov-config=pyproject.toml \
+        --cov-report=term \
+        --cov-report=xml \
+        --cov-append \
+        {posargs}
 
 [testenv:docs]
 description = Use sphinx to build the HTML docs.
